Flyttet kode til input komponenten
